% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_report.R
\name{AbundPcaPlot}
\alias{AbundPcaPlot}
\title{Function to generate PCA plot of abundance table : OTU,GENE,KO}
\usage{
AbundPcaPlot(abund.tab, sp.meta, meanCountCutOff = 1, groups = "ALL",
  title = "Title")
}
\arguments{
\item{abund.tab}{otu/gene table as matrix}

\item{sp.meta}{sample metadata as.matrix (ID,NICKNAME,CONDITION,SUBJECT,CATEGORY)}

\item{meanCountCutOff}{The mean count value a gene/otu must have to be considered for the plot.}

\item{groups}{list of selected biological condition for plot}
}
\value{
PCA plot
}
\description{
Function to generate PCA plot of abundance table : OTU,GENE,KO
}
\examples{
AbundPcaPlot(OTUDATA$otu.tab,OTUDATA$sp.meta,groups=c('ain','protein','cholic'))
AbundPcaPlot(OTUDATA$otu.tab,OTUDATA$sp.meta,groups=c('protein','cholic'))
AbundPcaPlot(KODATA$ko.tab,OTUDATA$sp.meta,groups=c('protein','cholic'))
}
