% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_report.R
\name{AbundPlotHeatmap}
\alias{AbundPlotHeatmap}
\title{Function to create a heatmap  of abundance data (ROWs:OTU,GENE,KO, COLS: SAMPLE NAMES)}
\usage{
AbundPlotHeatmap(abund.tab, sp.meta, meanCountCutOff = 1,
  groups = "ALL", fontsize_row = 5, fontsize_col = 5)
}
\arguments{
\item{abund.tab}{abundance table as.matrix}

\item{sp.meta}{sample metadata as.matrix (ID,NICKNAME,CONDITION,SUBJECT,CATEGORY)}

\item{groups}{list of biological condition(Healthy vs Disease)}
}
\value{
Heatmap plot
}
\description{
Function to create a heatmap  of abundance data (ROWs:OTU,GENE,KO, COLS: SAMPLE NAMES)
}
\examples{
OtuPlotHeatmap(OTUDATA$otu.tab,OTUDATA$sp.meta, meanCountCutOff = 50, groups=c('protein','cholic'),fontsize_row=5,fontsize_col=5)
}
