% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_report.R
\name{readingOtuData}
\alias{readingOtuData}
\title{Function to parse the content of OTU and KO files}
\usage{
readingOtuData(otutable_path, otutaxa_path, spmeta_path)
}
\arguments{
\item{otutable_path}{The file path of otu}

\item{otutaxa_path}{The file path of rdp classification}
}
\value{
OTU oject including otu tab
}
\description{
Function to parse the content of OTU and KO files
}
\examples{
readingOtuData("C:/Microbiome/MMDRPipeline/data/ben1/ben1_16s_otu_table.tsv","C:/Microbiome/MMDRPipeline/data/ben1/ben1_16s_otu_taxa.tsv","C:/Microbiome/MMDRPipeline/data/ben1/ben1_16s_sp_metadata.tsv")
}
