% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_report.R
\name{richness_phyloseq}
\alias{richness_phyloseq}
\title{Function for get richness and count summary for each sampleid using phyloseq}
\usage{
richness_phyloseq(abund.tab)
}
\arguments{
\item{abund.tab}{Integer abundance table including only interger values (row names are OTU/KO/GENES)}
}
\value{
richness summary table
}
\description{
Function for get richness and count summary for each sampleid using phyloseq
}
\examples{
otu_richness=richness_phyloseq(OTUDATA$otu.tab)
ko_richness=richness_phyloseq(round(KODATA$ko.tab))
}
