% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_report.R
\name{SampleStatSummary}
\alias{SampleStatSummary}
\title{Function for statistical summary of Abundance Table (OTU,GENE,KO,e tc) using phyloseq}
\usage{
SampleStatSummary(abund.tab)
}
\arguments{
\item{abund.tab}{otu table object from read.csv/table("otu.tsv"))}
}
\value{
Statistical summary  for each sample-id
}
\description{
Function for statistical summary of Abundance Table (OTU,GENE,KO,e tc) using phyloseq
}
\examples{
otu_sp_stat=SampleStatSummary(OTUDATA$otu.tab)
ko_sp_stat=SampleStatSummary(round(KODATA$ko.tab))
}
